+++ src/utils/terminal.c
@@ -4,7 +4,7 @@ void clearLine() {
     printf("\033[F\033[K");
 }
 
-void clearLines(unsigned i) {
+void clearLines(short i) {
     while(i > 0) {
         printf("\033[F\033[K");
         i--;
